Create separated projects and add teams name to the final result
