+++ Backend/RealManager/RealManager.Domain/Event.cs
