+++ Backend/RealManager/RealManager.Repositories/Interfaces/IPlayerRepository.cs
