+++ Backend/RealManager/RealManager.Repositories/Interfaces/IUserRepository.cs
