+++ Backend/RealManager/RealManager.Services/Interfaces/IPlayerService.cs
