+++ Backend/RealManager/RealManager.Services/Interfaces/IUserService.cs
