+++ Backend/RealManager/RealManager/RealManager.csproj
@@ -12,11 +12,13 @@
     </PackageReference>
     <PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="3.0.0" />
     <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="3.0.0" />
-    <PackageReference Include="MongoDB.Driver" Version="2.10.0" />
     <PackageReference Include="Swashbuckle.AspNetCore" Version="5.0.0-rc4" />
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\RealManager.Domain\RealManager.Domain.csproj" />
+    <ProjectReference Include="..\RealManager.Repositories\RealManager.Repositories.csproj" />
+    <ProjectReference Include="..\RealManager.Services\RealManager.Services.csproj" />
     <ProjectReference Include="..\RealManager.Shared\RealManager.Shared.csproj" />
   </ItemGroup>
