--- Backend/RealManager/RealManager/Repositories/Interfaces/IMongoRepository.cs
@@ -1,8 +1,0 @@
-namespace RealManager.Repositories.Interfaces
-{
-    public interface IMongoRepository
-    {
-        string ConnectionString{ get;set; }
-        string DatabaseName { get;set; }
-    }
-}
